Added NewField
This commit adds NewField to Foo, and also adds a working test to
TestEncode. But here we can see the problem. New contributors might
change the code and all tests are passing. But in fact, the code is
broken, because Decode is unable to reconstruct the Foo object.

In this example it might be obvious, that Decode must be changed as
well, but in large code bases this is not true.