Combine TestEncode and -Decode to TestEncodeDecode
Be kind to yourself and fellow contributors and always Test
Encode/Decode or Marshal/Unmarshal in one go.

When you see the both functions Encode and Decode each as its own "unit
under test", this might sound confusing. But in fact the unit under test
here must be the combination of both. One doesn't make sense without the
other.

When you run this commit, the tests won't pass. This is a great
improvement! No accident feature in Encode to break Decode. You will
encounter this problem very early, worst case in the CI. But we don't
depend on a sharp minded reviewer to gate this change.

Note: I don't want to say, that you must not necessarily start with
TDDing those functions with the combined test. You are happily allowed
to start with TestEncode and TestDecode to implement both functions. But
before you submit your work, do yourself the favor and replace both test
functions by TestEncodeDecode.

Just to say it out loud: You have to remove (test)code!!! But if you
don't you might not gain anything.